refactor: move epoch_sync API to EpochManagerAdapter
cc #6910
